+++ django_webix_sender/templatetags/send_methods_utils.py
@@ -2,10 +2,24 @@
 
 from django import template
 from django.conf import settings
+from django.contrib.auth.models import Group
 
 register = template.Library()
 
 CONF = getattr(settings, "WEBIX_SENDER", None)
+
+
+@register.simple_tag(takes_context=True)
+def user_can_send(context):
+    """ Returns boolean to indicate if user has send permission """
+
+    request = context['request']
+    if request.user.is_anonymous:
+        return False
+    if "groups_can_send" in CONF and \
+        not request.user.groups.intersection(Group.objects.filter(name__in=CONF['groups_can_send'])).exists():
+        return False
+    return True
 
 
 @register.simple_tag(takes_context=True)